legal_chatbot/requirements.txt
PyPDF2
pinecone
python-dotenv
sentence-transformers
tqdm
huggingface-hub
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
langchain-pinecone
pypdf
streamlit
chainlit
flask
fastapi
uvicorn
beautifulsoup4
requests